+++ core-ng/src/main/java/core/framework/impl/web/route/Route.java
@@ -4,7 +4,7 @@ import core.framework.api.http.HTTPMethod;
 import core.framework.api.log.ActionLogContext;
 import core.framework.api.util.Maps;
 import core.framework.api.web.exception.NotFoundException;
-import core.framework.impl.web.ControllerProxy;
+import core.framework.impl.web.ControllerHolder;
 import core.framework.impl.web.PathParams;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,9 +21,10 @@ public final class Route {
     private final Map<String, URLHandler> staticHandlers = Maps.newHashMap();
     private final PathNode dynamicRoot = new PathNode();
 
-    public void add(HTTPMethod method, String path, ControllerProxy proxy) {
+    public void add(HTTPMethod method, String path, ControllerHolder controller) {
         logger.info("route, {} {}", method, path);
         validator.validate(path);
+        controller.action = new ActionInfo(method, path).action();
 
         URLHandler handler;
         if (path.contains("/:")) {
@@ -35,10 +36,10 @@ public final class Route {
                 staticHandlers.put(path, handler);
             }
         }
-        handler.put(method, proxy);
+        handler.put(method, controller);
     }
 
-    public ControllerProxy get(String path, HTTPMethod method, PathParams pathParams) {
+    public ControllerHolder get(String path, HTTPMethod method, PathParams pathParams) {
         URLHandler handler = staticHandlers.get(path);
         if (handler == null) handler = dynamicRoot.find(path, pathParams);
         if (handler == null) throw new NotFoundException("not found, path=" + path);