+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.3.7
 * web: fix webservice controller inspection (getAnnotation/methodInfo)
+* mongo: find supports orderby/skip/limit
 
 ### 3.3.6 (8/7/2015 - 8/10/2015)
 * log: push action log to rabbitmq, index by ES/Kibana